Implement Tournament Tree deletion
